+++ src/reanimated2/animation/decay/index.ts
@@ -1,2 +1,2 @@
 export { withDecay } from './decay';
-export type { DecayAnimation, WithDecayConfig } from './decay';
+export type { DecayAnimation, WithDecayConfig } from './utils';
